chore: rename app folder to server folder
