+++ src/server.js
@@ -1,5 +1,5 @@
 const { createApp } = require('glip-integration-js');
 
-const appConf = require('./app/index.js');
+const appConf = require('./server/index.js');
 
 exports.server = createApp(appConf);
