+++ src/server/adaptiveCards/authToken.json
