+++ src/server/adaptiveCards/release.json
