+++ src/server/utils/findItemInAdaptiveCard.js
