+++ src/server/utils/formatAdaptiveCardMessage.js
