+++ src/server/utils/formatBugsnagMessage.js
